requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.5
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf==3.17.1
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2